third_party/WebKit/LayoutTests/shadow-dom/crashes/detach-node-call-recalc-assignment-crash.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<summary><div id=first style="display: list-item"></div><div id=second></div></summary>
<script>
test(() => {
  const summary = document.querySelector('summary');
  document.body.offsetLeft;
  document.querySelector('#second').remove();
  summary.remove();
  assert_true(true, "This test should not crash.");
}, "assigned_nodes should not be modified while in iterating it. See http://crbug.com/843069#c8");
</script>
